Move Util Functions
