+++ roscompile/src/roscompile/util.py
@@ -1,5 +1,6 @@
 import collections
 import os
+import re
 import stat
 
 import rospkg
@@ -53,3 +54,13 @@ def get_config():
         else:
             CONFIG = {}
     return CONFIG
+
+
+def convert_to_underscore_notation(name):
+    # https://stackoverflow.com/questions/1175208/elegant-python-function-to-convert-camelcase-to-snake-case
+    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
+    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
+
+
+def convert_to_caps_notation(name):
+    return ''.join([x.title() for x in name.split('_')])
